move BorderAlgo to own file, use enum only for config Conflicts: 	src/com/massivecraft/factions/entity/UConf.java
